SPL Engine (4.0-dev/5.0-dev): Reworked focus to studio and play after connecting script, plus a huge limitation is gonecd stationPlaylist/!
The scripts for toggling focus to studio and play after connecting has been modified to take the active encoder into account. This is same as the older lgobal plugin method.
Also, thanks to the reworked connection routine, broadcasters are no longer required to stay in encoder window until connection is established (a huge problem that plagued scripts such as that of NvDA).