Project Starfish (6.0-dev): Starfish gains ability to make of copy of itself
Thanks to code from NvDA core's config profiles dialogs, it is now possible to create new profiles. This can be done in two ways: branc new or as a copy of a base profile. In reality, the same dialog is used, with the copy flag used to control title text and create/copy operations. Because SPLCOnfig is a reference to one of the profiles, the copy function uses a new file instead of just appending an existing profile.
Also fixed an issue where SPLConfigPool would not be initialized when SPL is run again. This was due to the fact that SPLConfigPool is assigned None when the app module exits.