Track Tally (6.0-dev): Given a range of track indecies, calculate total running time.
In order for this to be done, announceTime function was split into two functions. For backwards compatibility, announceTime will now output human-readable time string coming from the new milliseconds to time (ms2time) function.
The track tally output isn't polished yet, but we now have enough foundation to tweak the output a little. This will be available in 6.0 due to API change.