Broadcast Profiles (6.0-dev): Never again pop up multiple config errors, plus broadcast profile will now be focused when config dialog is opened.
Due to the fact that 5.x and earlier was concerned with what's now called normal profile, the config error dialog was opened multiple times, one per invalid profile. This has been corrected - once a profile is found to have errors, the error code will be stored in the load status map to be retrieved by the init function. Also, the output of the config error dialog is such that each line will be of the format 'config name: error message'.
Also broadcast profile combo box will now be focused when config dialog is opened.