Cache Studio window handle for better performance.
Until now, it was necessary to call user32.FindWindow each time Studio handle was to be used in a routine. This is no longer the case: once the Studio window handle is found, it'l;l be cached so other routines cna look this up. As part of this change, a thread will be run at construction time to locate the handle and cache it once it is found.