Selective column announcements (6.0-dev): Initial work and debug code for selective column announcement.
When told to announce track columns differently, simulate fetching obj.description by building the column texts one piece at a time. This allows NVDA to announce only needed columns.
Because of some anoying limitations with COnfigObj, manually convert between strings and lists. This is done to convert column headers between ConfigObj and add-on formats and will be done when the config is loaded and saved.
Due to work involved and with GUI changes planned, this will be available in add-on 6.0.