Misc module (6.0-dev): Column lookup code reorganization
Sometimes, NvDA says the object is not a track item when in fact it is. This is possibly due to foreground window having wrong window class name. To fix this and to reorganize the module a bit, column content getter has been moved to the misc module.
In the app module, Track Dial wil still use obj._getCOlumnContent for backward compatibility reason. This is just a stub for calling splmisc version.
In Track Finder, when column search is active, use splmisc's version of column content getter to obtain the column text. This fixes a bug where searching for the same text in the same or different column results in error dialog being shown when in fact one of the tracks may contain column text as part of its description.
Due to extensive modifications, this will be exclusive to add-on 6.0.