Instant profile switch (6.0-dev): Initial work to support switching profiles.
When told by the broadcaster, switch to a predefined profile. This is done through a new command in SPL Assistant (F11, to be made assignable). NVDA will first ensure that the profile one wishes to use isn't the active profile before switching to the new profile (swaps config dictionary).
For now, a constant has been provided for testing purposes. In the future, a new button will be added to set selected profile as switch profile (except normal profile). More goodies to come later.
As this builds on top of broadcast profile management, this will be exclusive to 6.0.