Instant profile switching (6.0-dev): Fix a showstopper bug where switch button shoed wrong label when switch profile is undefined.
When displaying the label for the switch button, just check switch profile. Also, when switch profile is off, be sure to nullify prev profile, otherwise NVDA will assume one can switch back to the predefined prev profile when it can't.
With this, switch profile is ready for testing. The ability to save this across sessions will not come to life for a while.