Cancel event fixes.
If the active profile or switch profile is deleted or renamed, make sure to update the switch profile. In case the active profile is gone, set SPLConfig to normal profile, otherwise value error is raised as obsolete SPLConfig will be loaded.
To help with this change, two sanity check flags are introduced to handle renames and deletions. Hopefully no more need to revisit onCancel event unless significant bugs emerge.