Track Tool optimization (6.0-dev): Get column content improvements.
First of a series of optimizations to use SysListView32. For the first step, imported _getColumnContent from splstudio.splmisc module to reduce code duplication.
Individual column scripts now uses getColunContent. NOt all messages from previous add-on releases were ported to the new routine and performance might be a bit slower than expected (no caching yet). As of this commit, column scripts are working as expected, with more optimizations to come later (this is destined for add-on 6.0).