Broadcast profiles (6.0-dev): Significant API changes to support sorting.
In order to display sorted profile names, it was necessary to perform a surgery on splconfig module. To account for a case where config pool isn't sorted (especially after adding profiles), a dedicated checks and sorting routine (technically a call to sorted) is introduced as part of config dialog's constructor routines.

In addition, splconfig gained three new functions:
* getProfileIndexByName: Retrieve the index for the profile with a given name.
* getProfileByName: Retrieve the profile assocated with a given name.
* isConfigPoolSorted: Check whether profile names are sorted in the pool itself.

Also took this time to rewrite various event callbacks in the config dialog to account for the new API's and resulting changes (no more reliance on selection index alone, now string will also be considered via wx.Choice.GetStringSelection function).
This is destined for 6.0 (starting with the September snapshot).