Alarm UI message (6.6-dev): Initial foundation to allow alarm notification to be spoken and/or brailled and the GUI for it.
Currently NVDA only plays beep when end of track or song intro alarm kicks in. In 6.0, this will be extended to let NVDA announce messages. To handle this, a new setting called AlarmAnnounce will be introduced, along with a combo box to choose alarm notification (beep only, message only or both).