Disk I/O: small yet significant optimizations.
The default config and its validator partner is now a private data structure of the config module. Among other things, this removes duplicate code and to check vlaues against when saving confi database.
Also, when cpying a profile, do noot copy the file. Instead, an in-memory copy will be performed, with disk writes reserved for save config function.
There might be a possibility of backporting this to add-on 5.x.