Track item (6.0-dev): No more reliance on foreground window, window style will be consulted instead.
Previously, foreground window was checked. A major problem was that NVDA would not recognize that it is dealing with an actual track item (it'll just say it is an IAccessible item). This was more so after dismissing nVDA's own dialogs (alarm dialogs, for instance). To solve this problem, window style will be consulted (they don't change even across sessions).
For now Studio 5.10's track item window style was added. A possible future commit will add Studio 5.0x track items if their window style is different. As this change is a fundamental change, this will be used in add-on 6.0.