Place marker (6.0-dev): INitial work on placing place markers.
The place marker is used to quickly focus to a track. This is done by storing the actual object (not a filename, not the index). The set/find routine is part of SPL Assistant, and delete track has been modified to remove place marker object if the track just deleted was the place marker track.
In some cases, if a track above the place marker track is removed, a zombie object will be left behind in the place marker cache. In this case, obj.name and obj.description are None, and if this happens, the cached object will be removed.
So far, only one track will be marked. This will be available in add-on 6.0.