Place marker (6.0-dev): Few deeper changes.
Because of a bug where the place marker object will not update its position, a dedicated tuple will now be used to store place marker information. This tuple and place marker functions will be part of a new section in the app module code.
Also, place marker will make sure one is in a general track or a spot - place marker will not work for hour markers, break notes and others.
Some significant rearchitecting in Track Finder is in order (a separate branch will be used, as it affects Track Finder version 1 and the new place marker feature).