Place markers and column routines (6.0-dev): Column header fetcher changes for speedup.
If a large number of tracks have been added, column routines such as place marker setter and Track Dial becomes slow. To mitigate this, cache column header object in the app module. This is done whenever column routines are first called. In the end, the column headers should be cached automatically when the app module starts, but that's an optimization that'll be done later.