Window style check (6.0-dev): Perform bitwise and or check the modulo of the subtraction to distinguish track items from other list view items.
Using lists of known styles is fine if we know all possible style bits. However, this has a flaw in that it'll break if new track item styles are encountered. Besides, doing a bitwise and says track items have unique bit set (0x100000, a scrollable window), thus bitwise and is more reliable. However, for STudio 5.0x, do subtraction to really make sure we end up with either 0 or 0x100000, as the default style bit sets 0x100000 to zero.
This new routine should be future-proof when it comes to detecting track items reliably. However, keep an eye on style changes in future Studio versions.