Metadata streaming (6.0-dev): Announce a reminder message and play a sound if metadata URL's are configured.
The handle finder is now responsible for reminding the broadcaster to enable metadata streaming if such URL's are configured. Also, just in case the reminder announcement is misse,d an alarm wave file (from PaC Mate original) will be played to grab attention.
Added some new SPL Assistant commands related to metadata streaming, including numbers 1 through 4 for each URL and E to announce if URL's are configured.
This will be available in add-on 6.0 (requested by a long time beta tester).