Metadata streaming (6.0-dev): New UI and a new bollean list to store reminder value for each URL.
A new dialog to configure metadata streaming for each URL, which will be the basis for further work.
A new boolean list in the confspec will be used to store reminder flag for each URL. This will be used during startup to remind users to enable metadata streaming with the correct information.