Metadata streaming (6.0-dev): Changed UI, metadata reminder is ready for use by instant profiles.
Because instant profile switching may demand metadata reminder to be announcer, changed the UI for configuring this option to a combo box. By default, startup is selected. One can also choose to hear the reminder if switching to the online (instant switch) profile.
When switching profiles, check if metadata reminder option is set to instant. Same is done when the app module loads, except the value to be checked is startup. Lastly, when using instant value, pauses should not be there and speech should not be canceled (a new pause kwarg will ensure it).