Broadcast profiles (6.0-dev): Reset settings will reset all profiles.
A new function is available for resetting all profiles. Hopefully this should be called only when the user intentionally asks for it (add-on settings).
Also, due to reference issues, a reset flag will be stored in the normal profile when the profiles are all reset to defaults, to be removed just before the normal profile is saved. Because of this change, when saving profiles, normal profile will be saved last.
Hopefully no more need to touch broadcast profiles for a while until a rare bug is found by users.