Message verbosity (6.0-dev): Initial work on message verbosity.
It is now possible to specify how much message should be announced by NVDA. For now, there are two modes: beginner and advanced. For now, toggle messages will be affected by this change, with add-on messages to be fetched from a named tuple later.
Also added the user interface to configure message verbosity.
This will be part of add-on 6.0.