Time range finder (6.0-dev): A new dialog to search for a track with duration faling between minimum and maximum.
A variation of track finder dialog is introduced to search for track with a duration that falls in between min and max. This is done by two contorl groups, each containing two spin controls for minute and second. Once OK is clicked, Studio API is used to obtain track duration (for this reason, time range finder takes two additional parameters, namely the focused object and the studio API function wrapper).
This will be exclusive to add-on 6.0 (suggested by a long time beta tester).