Message verbosity (6.0-dev): Laid the foundation for message verbosity.
A new messages pool has been added to splconfig module (to be split into its own module later). This is a multimap, consisting of message category, value and a tuple of messages according to verbosity level. For now, status announcement script uses this new method as a proof of concept (more to come later).