Number entry fields: Values are now highlighted, microphone alarm field is now a spin control (number entry dialog in app module portion), now can be set up to two hours (7200 seconds).
With the introduction of spin controls, there is no longer a need to perform integer to string conversion and back in microphone alarm dialog. In 6.0, mic alarm dialog is still a separate dialog from other alarm dialogs, but it is now a number entry dialog. There might be a possibility to incorporate mic alarm dialog into the common alarm dialog, but it will require some changes to common alarm dialog API (will not be done for 6.0, as public beta cycle is near).
Also microphone alarm now has a maximum value - 7200 seconds, or two hours (suggested by Jerry Mader).