6.0 prebeta: Fix for a problem where double focus event was fired when moving to a different track after searching for one.
Partly reverts a commit that removed the focus checker from 0.01 (now reinstated), as removing this caused focus to be fired on a previously focused item after finding a track and moving to a different track. The issue of items firing focus events will be dealt with in a future version.