Rainbow/red (7.0 prealpha): A new config format has arrived.
Starting with add-on 7.0, add-on config will be organized into sections. For now, added a method to convert old-style config to the new format, and next few commits will gradually transition the current config routines to the new format. Old style config will be retained (but will not be used) in add-on 7.0, to be removed in 7.2/8.0-dev next year.
Due to this change, the new format is destined for 7.0, while the old format was retained for backward compatibility.