Rainbow/red (7.0 prealpha): Config unlock and presave now uses new sectionized config format.
Because config are stored in sections, it is important to make sure errors are preserved when validating the config file. This also allows the unlock routine to restore a faulty setting stored in a subkey instead of restoring the entire section. Also, make sure the subkey passes (int is returned).
In presave, when deleting profile-sepcific sections, be sure the section is empty before proceeding (slight modification to 6.0's presave routine).
For now, UI doesn't work as runtime application of these settings are not implemented (will come later). Because of extensive changes, this is destined for 7.0.