6.0 beta 3: Add ability to apply metadata stremaing toggle to the currently active profile.
At the moment the real way to make metadata streaming toggle become part of the active profile is to configure it via add-on settings dialog. The quick toggle dialog is meant to toggle streaming on the fly, but there might be times when it is desirable to make it permanent (and somehow causing a bug where NVDA is stuck to using metadata enabled vlaues from saved profiles, reported by Roger Stewart). To make this happen, a new checkbox has been added in quick toggle version of the metadata stremaing dialog to apply current toggle to the active profile (intentionally false).