5.6: NVDA-specific hack to announce playing track automatically.
According to Brian Hartgen (JAWS script writer), auto scroll in studio might be broken (subsequently confirmed). Thus NVDA will now announce currently playing track automatically (a hack). Once this is fixed in Studio, this hack will be modified to operate only when Studio is not in focus.
This is a critical issue for 5.x and 6.0.