6.0: Fix a critical issue related to column inclusion set.
If the profile is reset to defaults, included columns key becomes a list. Thus manual conversion is required.
Also, when changing included columns set and then clicks cancel button, the new set is retained. To prevent this, when cancel button is clicked from add-on settings, clear the current included columns set.