Auto update check (7.0-dev): Initial work on chekcing for add-on updates automatically.
A new key (PDT) will store when the update check happened (PDT = package date). Also, a new config key has been added to control auto update check behavior (to be made configurable under advanced options later).
When the app module starts, if NvDA is told to check for add-on updates, update will be checked. For now, a sign-in message will be hard.
The update check routine in splupdate module has been updated to support auto update (if this is on, no progress or error UI will be shown just like NVDA Core).
As this is an improvement over project gold (update check foundation), this will be exclusive to add-on 7.0.