Auto update check (7.0-dev): Better update qualifier check.
In case one is running a newer version, close the loophole where one can 'upgrade' to an older version of the add-on. Also normalize the version string to match the one sent by the server.
Fixed a problem where add-on said there was an update when there isn't (how did an exclamation mark ended up there?).