Auto update check (7.0-dev): Automatic update check is now available.
Added a new timer which calls the update check routine. Due to limitations, a proxy is used (the timer constructor does not allow arguments).
The update checker gains a new keyword argument that takes a copy of the auto update key from add-on configuration. This is used to tell if the timer should resume. Also, regardless of check type, the check time will always be updated whenever the update check happens (thus, PDT is now package update time).
So far, auto update is working (will be checked every 24 hours). The next step is disabling the auto checker in specific cases such as when an instant switch profile is active.