6.1: Fixed a possible double announcement of metadata streaming if using an expired copy of Studio demo.
When the demo expires and if the app module starts, the Studio handle finder will run forever. This causes the following:
1. User reinstalls Studio after dismissing the expiration screen.
2. The app module reloads. This causes a second instance of the handle finder thread to come to life.
3. Once both threads locate the Studio handle and if metadata streaming option is set to announce at startup, the information will be announced twice (a queuing problem).
To fix this, an event (threading.Event) will be used to tell the handle finder to stop looking for the handle when Studio app module terminates. This warrants a readme entry in add-on 6.1.