Auto update check (7.0-dev): If a socket error occurs, retry after a while.
If an Internet connection is lost, urllib throws IOError (tecnically, a socket error). Ifthis happens, the add-on will wait ten minutes before trying again, and in the process, set a flag reminding itself to check for update after a failure. Once this flag is set, update check time will not be set until a successful connection is made.