6.1: Allow SPL Assistant commands with gestures defined to work in other Studio windows.
If gestures are assigned to some SPL Assistant commands, it is no longer required for user to be located on the Studio main window. This is done by using a trick from global plugin: using IAccessible's getNVDAObjectFromEvent function (used when attempting to focus to sTudio window from other programs). The try block remains for backward compatibility (might be removed in add-on 7.0 if this feature works as advertised).