6.1: Changed the wording of config error dialog if the only broken configuration is the normal profile.
If the only broken configuration is normal profile, display a different text when config error dialog is shown (suggested by Duyahn Walker and subsequently confirmed). Also added additional translatable strings (this is an exception, as maintenance mode does not allow this, but because it affects user experience, this was granted for this commit). This will be included in add-on 6.1.