Time-based profile switching (7.0-dev): Initial work on tim-based profile switching/triggers.
A number of new functions and other support routines added to support time-based profile switching. In order to differentiate this from the main settings dictionary, profile triggers will be stored as a pickle (data dumps), which allows needed data structures to be reconstructed.
Added a triggers dialog in main settings dialog to allow specifying dates and times on which a given profile should be triggered. For now, the actual switching code isn't there (that's for later).
This will be one of the highlights of add-on 7.0.