Profile triggers (7.0-dev): Fixed next profile setter and dictionary lookup issue.
Added 'time-based profile' flag for profiles. This flag will be set when a trigger is found.
Fixed an issue where a weekly show returned None for profile trigger spec if the trigger was set on a date other than the broadcasting date. Also fixed an issue where clearing triggers did not remove the associated profile from triggers dictionary.