Time-based profile switching (7.0-dev): Huge advancements, timer, triggers pickle and support routines.
With the triggers spec defined, a pickle (data persistence) dictionary is now used to store profile triggers. This pickle is loaded when the initConfig is called, used throughout runtime and dumped away during saveConfig.
A number of support routines are now available for managing profile triggers. A notable function is triggerProfileSwitch which is essentially the trigger version of instant switch routine.
A new timer class (SPLCountdownTimer) is now available to announce countdown time and to run a specified function. This is used for profile triggers only.
Thanks to the new timer functionality, profile triggers is now live (configurable from triggers dialog in add-on settings). This will be available in add-on 7.0.