6.2: Playlist remainder announcement config UI.
Somehow, when SPL Assistant, D is pressed, the remaining time for the current hour is announced contrary to the description of this command (reported by Roger Stewart, confirmed by Joseph Lee). Because of this, a check will be done in playlist remainder command to see if the remainder of the entire playlist should be announced. As a first step, a new combo box in advanced settings will allow a user to configure this check.
This is destined for add-on 6.2 (it'll be included in 7.0 after modifications including selecting the best section to put this value in).