6.2: Completely changed playlist remainder calculation strategy.
According to SPL Studio developer, the duration shown in the playlist viewer is segue, which explains why the trakc time analysis method didn't work. To get around this, a variant of column search will be used - searching for duration, converting the string output into seconds, then displaying this in a friendly way.
The time announcer has gained a new boolean flag to treat time in milliseconds or not (default is true). For add-on 6.x, playlist remainder will call time announcer with this flag set to true for backward compatibility, but in 7.0, milliseconds flag will be turned off.
If this works as advertised, this will be included in add-on 6.2 with a caveat: will need to add a warning about possible inaccuracies in the readme.