Category sounds (7.0-dev): Initial implementation of category sounds.
Category sounds is used to alert the broadcaster as to what type of track one is focused on. This is done by looking up category column and consulting a tones dictionary (might be replaced with a wave file later if requested). For now, this option is enabled and will be disabled in production releases so users can configure this (this is exclusive to add-on 7.0 due to the new config format).