Playlist remainder and friends (7.0-dev): Major change to playlist remainder and other commands in SPL Assistant.
Due to changes from add-on 6.2, a number of SPL Assistant commands are reassigned and one new command is now introduced. Specifically:
* Shift+H is now Shift+S.
* D: Split into two functions, with the total track duration counter retained.
* Shift+H: Split from playlit remainder (D), now announces what SPL Assistant, D used to announce.
As a result, playlit duration key is no longer used in SPLConfig. This key (stored under Advanced section) is deprecated and will be removed in 2016 (as early as add-on 7.2, and 8.0 will fully retire this key).
This refactor is destined for add-on 7.0 and later.