Update check (7.0-dev): Split update check routine.
In order to properly implement config caching, it was necessary to split update check routines and storage. splupdate now includes initialize and termiante methods to load and store update size and check date, previously performed by config manager. This frees splstudio.ini to store just add-on settings.
This refactoring is destined for add-on 7.0 and will take effect after this commit.