Sectionized validator (7.0-dev): Unlock config now checks profile-specific sections when unlocking broadcast profiles.
NVDA Core provides a way to validate only sections, using an undocumented section keyword argument in configobj.validate. A simpler version will be used in this add-on: using profile-specific confspec and using it when unlocking profiles other than normal profile.

This change means:
* No need to call discard global settings anymore as broadcast profiles will have just profile-specific sections.
* When some keys fail validation checks, the reset routine can work on all profiles instead of checking to see which profile we're dealing with.
There is slight performance improvement with this approach. If this approach is successful, it'll be included in add-on 7.0.