Config cache (7.0-dev): New profiles are now saved to disk.
Oops, a downside to caching found: if a new profile is created (brand new or as a copy), this fact isn't recorded. The cache will see the 'new' profile as an existing one, therefore it'll not save the newly created profile to disk.
Also, when copying profiles, profile-specific keys are copied insead of all keys being copied over.