Columns explorer (7.0-dev): Configuration UI added.
A new COlumns Explorer dialog has been added to add-on settings. This dialog houses six or ten combo boxes (six if using Studio 5.0x) used to configure which column should be 'explored' (looked up) when SPL Assistant, nubmer keys are pressed. For visual clarity, these combo boxes are laid out on two rows (first row are 5.0x, seocnd row is 5.1x).
This is destined for add-on 7.0.