Add-on settings dialog/instant profile swithcing (7.0-dev): instant switch toggle is now a checkbox, laid foundation for modifying profile flags.
For consistency and to improve user experience, instant switch toggle is now a checkbox. This allows easier way to modify profile flags.
In regards to profile flag modifications, a new function has been added to modify flags given the entry index, action to be performed and the flag in question. Instant switch toggle now uses this function, and a future commit will extend this to time-based triggers.
The new UI change is destined for add-on 7.0 and later.