Profile flags (7.0-dev): correct profile is selected after creating certain profiles.
Because config pool isn't sorted, using active index flag in SPLConfig has a major flaw: wrong profile will be selected. This occurs if a profile that starts with a letter that comes earlier than the first broadcast profile is created (for example, if broadcast profile 1 was named 'test' and a profile named 'acb' is created). This has been corrected by forcing selection index to come from profile names stored in the dialog.
This is specific to add-on 7.0.