6.3: Optimize a specific report focus case where screen order is off yet column order and included columns are same as defaults.
Previously, NVDA didn't case if column order and inclusion were set to default values. As a result, description pieces werebuilt when it was unnecessary, wating valuable computing time. This case is now caulght and handled.
To handle this subtle case, a new function (private) has been added to splconfig module which will tell NvDA to build description pieces when it is absolutely necessary.
This is destined for both 6.3 and 7.0 (7.0 porting requires sectionizing relevant settings and few more optimizations later).