6.3: When profiles are deleted, focus moves to the active profile (normal profile is active profile is gone).
Previously when deleting a profile, normal profile was selected as the default. This produced two issues:
1. If there was an active profile that has survived, focus will not go to that profile.
2. When cancel button is pressed, normal profile is activated no matter what kind of profile was deleted.
This has been fixed by checking to see if the active profile is still alive, and if so, the live active profile will be selected, otherwise normal profile will be used. This is destined for add-on 6.3 and later (affects 7.0 but requires few more tweaks).