6.3: Remove unnecessary statements, catch exception when attempting to rename profiles.
When attempting to rename a profile and if the file path does not exist, Windows will throw an error. Previsouly this check wasn't done, which caused renamed string to fail to show up (a major bug).
Also removed unnecessary statements including a print statement from 6.0.
This is destined for add-on 6.3 and later.