Add-on installer and config management (7.0-dev): Handle parsing error gracefully.
During add-on install, if parsing error occurs, NVDA will terminate add-on installation procedure. Thus skip failed config files so it can be handled at runtime.
During runtime, when config loads, catch parse error. If this ever happens, completely empty the offending file, then try loading again and record this fact so users will be made aware of this problem (doing so resets everything to defaults).
This is destined for 7.0 and later (no need to port this to 6.x as there is no config conversion routine).