Time-based profile switching (7.0-dev): Handle a subtle case where time-based profile could be deleted by accident by preventing deletion operation.
While broadcasting a show, it might not be desirable to let profiles be deleted, including the very profile (time-based one) the broadcaster is using. Thus while a show is broadcasting, deletion operation will not be allowed (once broadcasting is done, then deletion will be allowed).
This is a temporary workaround. A more elegant solution that allows broadcasters to delete currently active time-based profile will be added in future commits.