Time-based profile switching (7.0-dev): Fixed inconsistency seen when broadcaster switches to another profile in the midst of broadcasting a show.
If a time-based profile is active and if a new profile becomes active, the prev profile was nullified. This creates inconsistency between prev profile flag and trigger flag (trigger flag is on yet prev profile is null). To solve this, onOK handler checks to make sure trigger flag is off before nullifying the prev profile (this is done when switch profile stirng from add-on settings dialog is no more).
Some GUI commits to come later (mostly to finish time-based switching code and to borrow triggers dialog design from NVDA Core). In the meantime, the master branch will focus on tweaks and fixes/enhancements.