7.0-dev: Remove unnecessary reset variable and unused title string in unlock routine.
Thanks to use of a separate runtime dictionary, it is no longer necessary to store reset flag. Also, because error flags are stored into errors dictionary when profiles are unlocked, there's no need to keep the title for the error dialog around (besides, error dialog is no longer called from unlock routine). With this, another vestige of add-on 5.x-style config is removed.