Update check (7.0-dev): Remove old-style keys while caching normal profile.
Somehow, due to caching mechanism, old-style update keys were kept when settings did not change (should save returns false). Therefore the old-style update keys will be checked and removed before caching the normal profile. At the same time, the cache will be told that normal profile should be saved via a new flag.
This if for add-on 7.0 only (to be removed in 7.1).