Triggers UI (7.0-dev): Revamped triggers dialog (foundation laid).
The triggers dialog is now used to set various trigger flags, including instant switch profile and time-based switching. For now, the checkboxes were added. The old method of toggling instant switch flag from main add-on settings and selecting dates to enable time-based switching is still in effect, but this will change later.
This is destined for add-on 7.0.