6.3 (security): Prevent mishaps and memory leak due to encoder support flags adn maps not being cleaned up when app module is gone.
In theory, an app module isn't supposed to touch global plugin data structures when terminated. However, due to the fact that SPL app module controls SPL Utilities (a global plug-in), it is necessary to perform cleanup actions. Previsouly, the flags used for encoder support weren't deleted, which posed security risks and caused memory/resource leaks.
Specifically:
* A hacker could notice that the encoder is registered in sys.modules, thus one can modify flags set such as focusing to Studio to fool NVDA to monitor a different encoder or remove encoders outright.
* Not only this results in a security problem, it caused memory leaks, as the encoder flags would not be empty.
* An additional bug was found where if a broadcaster restarts Studio without restarting NVDA, encoder monitor count would be incremented forever (if not monitored carefully).
A new cleanup function is now available to close this hole by nullifying flags set and resetting encoder monitor count.
This affects add-on 5.0 and later, to be patched in add-on 6.3. Add-on 7.0 will see enhancements to this patch, namely thread management.