7.0 optimization: bitwise and before adding track item class to overlay classes list.
Previously, absolute values of track item style minus known style and anding it with 1048576 was sufficient. However, this meant other items such as requests were treated as track items (if there is demand for it, Track Dial might be extended to cover them, but that's for 8.0 if accepted). Thus a bitwise and against known style bits (0x100000, 0x8000, 0x1000) will be done to weed out track item pretenders.
This optimization may need further testing (will be included if accepted).