Alarm dialogs: A new function just for handling GUI routines.
What used to be two separate GUI handling routines in alarm dialogs, these are now part of a brand new function with the scripts calling this function with just needed arguments. Not only this makes code more readable, it reduces translatable string duplication among other savings.
This is destined for add-on 7.0 (with this commit, major optimizations for the app module is complete).