Broadcast profiles and cache (7.0-dev): Do not run presave before a broadcast profile is marked for saving.
Suppose a broadcaster changed a setting in broadcast profiles, restarts NvDA or Studio and changes a setting from a broadcast profile back to default value. When this happens, the newly saved value will not be written to disk, resurrecting the old value next time the app module is run. Because of this, when saving profiles other than normal profile, shouldSave function will be run first, and if it should be saved, presave will be invoked against the profile in question (does not affect normal profile, as everything will be stored there).
This is destined for 7.0 (no need to document this).