Config UI (7.0-dev): SPLConfig UI dialogs are now part of a new module.
One of the items to be implemented in future add-on releases is a new add-on settings dialog interface, with a list of categories replacing various settings dialogs. Another topic under consideration is rewrite of config management routine to use a new config hub data structure similar to how NVDA Core manages configuration and profiles. To facilitate this, a new splconfui module houses basic config dialogs that were part of the main SPL config module.
At the moment the config modules are ready. This isn't ready for prime time yet due to ongoining optimization work and because the main app module file wasn't modified to use the new module. The latter will be done once optimization is finished.
This lays the foundation for 7.0 LTS and 8.0-dev work.