7.0 optimization/time-based profiles: INtegrate find next air date into set next profile function.
Because findNextAirDate is used by only one function, it makes sense to integrate this function into setNextTimedProfile function. This also allows the function to be refactored - instead of returning a list outright, it'll calculate date delta, and then will return the profile entry with the new date information applied (except if the current date is the broadcasting day and the sho hasn't started). Due to complexity, math.log will be kept (quite clever and readable).
This is destined for add-on 7.0 after some testing.