Track name announcement (7.0 optimization): Simplified track name object lookup.
Instead of looking for status bar, lookup title name object directly using status function (part of the app module since the begining). This was possible thanks to the fact that the path to the track title object was defined months ago (part of add-on 6.x). Also, because caching, it is faster to compare objects directly instead of hunting for nearby status objects and watching out for attribute error. As a result, the part of name change event handler that deals with track name announcement is much simpler.
This is destined for add-on 7.0 after user testing.